+++ HISTORY.md
@@ -1,6 +1,13 @@
 Release History
 ===============
 
+0.3 (2015-06-23)
+------------------
+
+* Partial account support
+* Better headers/metadata management
+* Reviewed listing features
+
 0.2 (2015-04-09)
 ------------------